Moved the ray tracing logic to a separate class and added some comments.
