+++ Assets/Scripts/Statistics/ExplorationTracker.cs
@@ -16,6 +16,7 @@ namespace Dora.Statistics
         private ExplorationVisualizer _explorationVisualizer;
         
         private SimulationMap<ExplorationCell> _explorationMap;
+        private RayTracingMap<ExplorationCell> _rayTracingMap;
         private readonly int _explorationMapWidth;
         private readonly int _explorationMapHeight;
 
@@ -28,30 +29,7 @@ namespace Dora.Statistics
             _explorationVisualizer = explorationVisualizer;
             _explorationMap = collisionMap.FMap(isCellSolid => new ExplorationCell(!isCellSolid));
             _explorationVisualizer.SetMap(_explorationMap, collisionMap.Scale, collisionMap.ScaledOffset);
-            //_explorationMap.Raytrace(new Vector2(-21.8845768f, 9.14773464f), 232, 15.0f, (index, cell) => true);
-            _explorationMap.Raytrace(new Vector2(6.42033148f, -34.4225616f), 228, 15.0f, (index, cell) => true);
-            
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
-            Vector2 pos = new Vector2(0.1f, 0.9f);
-            const int traces = 90 * 500;
-            for (int i = 0; i < traces; i++)
-            {
-                var angle = i * 4;
-                if (i * 2 % 45 == 0) continue;
-                    
-                _explorationMap.Raytrace(pos, angle, 15.0f, (index, cell) =>
-                {
-                    if (cell.isExplorable && !cell.IsExplored)
-                    {
-                        cell.IsExplored = true;
-                    }
-            
-                    return cell.isExplorable;
-                });
-            }
-            sw.Stop();
-            Debug.Log($"Execution time for {traces} traces: {sw.ElapsedMilliseconds}");
+            _rayTracingMap = new RayTracingMap<ExplorationCell>(_explorationMap);
         }
         
         public void LogicUpdate(SimulationConfiguration config, List<MonaRobot> robots)
@@ -66,14 +44,13 @@ namespace Dora.Statistics
                     var angle = i * 4;
                     if (i * 2 % 45 == 0) continue;
                     
-                    _explorationMap.Raytrace(robot.transform.position, angle, visibilityRange, (index, cell) =>
+                    _rayTracingMap.Raytrace(robot.transform.position, angle, visibilityRange, (index, cell) =>
                     {
                         if (cell.isExplorable && !cell.IsExplored)
                         {
                             cell.IsExplored = true;
                             newlyExploredTriangles.Add(index);
                         }
-
                         return cell.isExplorable;
                     });
                 }
